+++ conf/smsapi-example.inc.php
