+++ src/Controller/SetQuestionsController.php
@@ -19,9 +19,18 @@
 #
 #==============================================================================
 
-# This page is called to send a reset token by mail
+# This page is called to set answers for a user
 
-class SendTokenController extends Controller {
+namespace App\Controller;
+
+use App\Framework\Controller;
+use App\Framework\Request;
+
+use App\Service\LdapClient;
+use App\Service\RecaptchaService;
+use App\Service\UsernameValidityChecker;
+
+class SetQuestionsController extends Controller {
     /**
      * @param $request Request
      * @return string
@@ -36,16 +45,22 @@ class SendTokenController extends Controller {
 
     private function isFormSubmitted(Request $request) {
         return $request->get('login')
-            && ($this->config['mail_address_use_ldap'] || $request->request->get('mail'));
+            && $request->request->get("password")
+            && $request->request->get("question")
+            && $request->request->get("answer");
     }
 
     private function processFormData(Request $request) {
-        $login = $request->get('login');
-        $mail = $request->request->get("mail");
+        $login = $request->get("login");
+        $password = $request->request->get("password");;
+        $question = $request->request->get("question");
+        $answer = $request->request->get("answer");
 
         $result = '';
         if (empty($login)) { $result = "loginrequired"; }
-        if (!$this->config['mail_address_use_ldap'] and empty($mail)) { $result = "mailrequired"; }
+        if (empty($password)) { $result = "passwordrequired"; }
+        if (empty($question)) { $result = "questionrequired"; }
+        if (empty($answer)) { $result = "answerrequired"; }
         if($result != '') {
             return $this->renderFormWithError($result, $request);
         }
@@ -78,68 +93,42 @@ class SendTokenController extends Controller {
             return $this->renderFormWithError($result, $request);
         }
 
-        // Check mail
-        $result = $ldapClient->checkMail($login, $mail);
+        $context = array();
+
+        // Check password
+        $result = $ldapClient->checkOldPassword3($login, $password, $context);
         if($result != '') {
             return $this->renderFormWithError($result, $request);
         }
 
-        // Build and store token
-
-        // Use PHP session to register token
-        // We do not generate cookie
-        ini_set("session.use_cookies",0);
-        ini_set("session.use_only_cookies",1);
-
-        session_name("token");
-        session_start();
-        $_SESSION['login'] = $login;
-        $_SESSION['time']  = time();
-
-        if ( $this->config['crypt_tokens'] ) {
-            $token = encrypt(session_id(), $this->config['keyphrase']);
-        } else {
-            $token = session_id();
+        // Register answer
+        $result = $ldapClient->changeQuestion($context['user_dn'], $question, $answer);
+        if($result != 'answerchanged') {
+            return $this->renderFormWithError($result, $request);
         }
 
-        // Send token by mail
-        $reset_url = generate_reset_url($this->config['reset_url'], array('token' => $token));
-
-        if ( !empty($reset_request_log) ) {
-            error_log("Send reset URL $reset_url \n\n", 3, $reset_request_log);
-        } else {
-            error_log("Send reset URL $reset_url");
-        }
-
-        /** @var MailNotificationService $mailNotificationService */
-        $mailNotificationService = $this->get('mail_notification_service');
-        $data = array( "login" => $login, "mail" => $mail, "url" => $reset_url ) ;
-        $success = $mailNotificationService->send($mail, $this->config['messages']["resetsubject"], $this->config['messages']["resetmessage"].$this->config['mail_signature'], $data);
-
-        if($success) {
-            return $this->renderPageSuccess();
-        } else {
-            return $this->renderFormWithError("tokennotsent", $request);
-        }
+        return $this->renderPageSuccess();
     }
 
     private function renderFormEmpty(Request $request) {
-        return $this->render('sendtoken.twig', array(
-            'result' => 'emptysendtokenform',
-            'login' => $request->get('login'),
-        ));
+        return $this->renderForm('emptysetquestionsform', $request);
     }
 
     private function renderFormWithError($result, Request $request) {
-        return $this->render('sendtoken.twig', array(
+        return $this->renderForm($result, $request);
+    }
+
+    private function renderForm($result, Request $request) {
+        return $this->render('setquestions.twig', array(
             'result' => $result,
             'login' => $request->get('login'),
+            'questions' => $this->config['messages']["questions"],
         ));
     }
 
     private function renderPageSuccess() {
-        return $this->render('sendtoken.twig', array(
-            'result' => 'tokensent',
+        return $this->render('setquestions.twig', array(
+            'result' => 'answerchanged',
         ));
     }
 }
