+++ src/Service/MailNotificationService.php
@@ -19,14 +19,30 @@
 #
 #==============================================================================
 
-require_once __DIR__ . '/src/autoload.php';
+namespace App\Service;
 
-use App\Application;
-use App\Framework\Request;
+use App\Utils\MailSender;
 
-$app = new Application(__DIR__ . '/conf/config.inc.php');
+class MailNotificationService {
+    /** @var MailSender */
+    private $mailSender;
+    private $mail_from;
+    private $mail_from_name;
 
-$request = Request::createFromGlobals();
+    public function __construct($mailerSender, $mail_from, $mail_from_name)
+    {
+        $this->mailSender = $mailerSender;
+        $this->mail_from = $mail_from;
+        $this->mail_from_name = $mail_from_name;
+    }
 
-$response = $app->handle($request);
-$response->send();
+    public function send($mail, $subject, $body, $data) {
+        $success = $this->mailSender->send_mail($mail, $this->mail_from, $this->mail_from_name,$subject, $body, $data);
+
+        if(!$success) {
+            error_log("Error while sending email notification to $mail (user ${data['login']})");
+        }
+
+        return $success;
+    }
+}
