+++ src/Service/UsernameValidityChecker.php
@@ -1,0 +1,62 @@
+<?php
+#==============================================================================
+# LTB Self Service Password
+#
+# Copyright (C) 2009 Clement OUDOT
+# Copyright (C) 2009 LTB-project.org
+#
+# This program is free software; you can redistribute it and/or
+# modify it under the terms of the GNU General Public License
+# as published by the Free Software Foundation; either version 2
+# of the License, or (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# GPL License: http://www.gnu.org/licenses/gpl.txt
+#
+#==============================================================================
+
+namespace App\Service;
+
+/* @function string check_username_validity(string $username, string $login_forbidden_chars)
+ * Check the user name against a regex or internal ctype_alnum() call to make sure the username doesn't contain
+ * predetermined bad values, like an '*' can allow an attacker to 'test' to find valid usernames.
+ * @param username the user name to test against
+ * @param optional login_forbidden_chars invalid characters
+ * @return $result
+ */
+function check_username_validity($username,$login_forbidden_chars) {
+    $result = "";
+
+    if (!$login_forbidden_chars) {
+        if (!ctype_alnum($username)) {
+            $result = "badcredentials";
+            error_log("Non alphanumeric characters in username $username");
+        }
+    }
+    else {
+        preg_match_all("/[$login_forbidden_chars]/", $username, $forbidden_res);
+        if (count($forbidden_res[0])) {
+            $result = "badcredentials";
+            error_log("Illegal characters in username $username (list of forbidden characters: $login_forbidden_chars)");
+        }
+    }
+
+    return $result;
+}
+
+class UsernameValidityChecker {
+    private $login_forbidden_chars;
+
+    public function __construct($login_forbidden_chars)
+    {
+        $this->login_forbidden_chars = $login_forbidden_chars;
+    }
+
+    public function evaluate($login) {
+        return check_username_validity($login, $this->login_forbidden_chars);
+    }
+}
