+++ src/config_compat.php
@@ -1,0 +1,80 @@
+<?php
+#==============================================================================
+# LTB Self Service Password
+#
+# Copyright (C) 2009 Clement OUDOT
+# Copyright (C) 2009 LTB-project.org
+#
+# This program is free software; you can redistribute it and/or
+# modify it under the terms of the GNU General Public License
+# as published by the Free Software Foundation; either version 2
+# of the License, or (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# GPL License: http://www.gnu.org/licenses/gpl.txt
+#
+#==============================================================================
+
+#==============================================================================
+# PHP modules
+#==============================================================================
+# Init dependency check results variable
+$dependency_check_results = array();
+
+# Check PHP-LDAP presence
+if ( ! function_exists('ldap_connect') ) { $dependency_check_results[] = "nophpldap"; }
+else {
+    # Check ldap_modify_batch presence if AD mode and password change as user
+    if ( $ad_mode and $who_change_password === "user" and ! function_exists('ldap_modify_batch') ) { $dependency_check_results[] = "phpupgraderequired"; }
+}
+
+# Check PHP mhash presence if Samba mode active
+if ( $samba_mode and ! function_exists('hash') and ! function_exists('mhash') ) { $dependency_check_results[] = "nophpmhash"; }
+
+# Check PHP mbstring presence
+if ( ! function_exists('mb_internal_encoding') ) { $dependency_check_results[] = "nophpmbstring"; }
+
+# Check PHP xml presence
+if ( ! function_exists('utf8_decode') ) { $dependency_check_results[] = "nophpxml"; }
+
+# Check keyphrase setting
+if ( ( ( $use_tokens and $crypt_tokens ) or $use_sms ) and ( empty($keyphrase) or $keyphrase == "secret") ) { $dependency_check_results[] = "nokeyphrase"; }
+
+# Action
+
+if (!isset($default_action)) { $default_action = "change"; }
+
+#==============================================================================
+# Other default values
+#==============================================================================
+if (!isset($ldap_login_attribute)) { $ldap_login_attribute = "uid"; }
+if (!isset($ldap_fullname_attribute)) { $ldap_fullname_attribute = "cn"; }
+if (!isset($pwd_forbidden_chars)) { $pwd_forbidden_chars = ""; }
+if (!isset($hash_options)) { $hash_options = array(); }
+if (!isset($samba_options)) { $samba_options = array(); }
+if (!isset($ldap_starttls)) { $ldap_starttls = false; }
+
+# Password policy array
+$pwd_policy_config = array(
+    "pwd_show_policy"         => $pwd_show_policy,
+    "pwd_min_length"          => $pwd_min_length,
+    "pwd_max_length"          => $pwd_max_length,
+    "pwd_min_lower"           => $pwd_min_lower,
+    "pwd_min_upper"           => $pwd_min_upper,
+    "pwd_min_digit"           => $pwd_min_digit,
+    "pwd_min_special"         => $pwd_min_special,
+    "pwd_special_chars"       => $pwd_special_chars,
+    "pwd_forbidden_chars"     => $pwd_forbidden_chars,
+    "pwd_no_reuse"            => $pwd_no_reuse,
+    "pwd_diff_login"          => $pwd_diff_login,
+    "pwd_complexity"          => $pwd_complexity
+);
+
+if (!isset($pwd_show_policy_pos)) { $pwd_show_policy_pos = "above"; }
+
+if (!isset($messages)) { $messages = array(); }
+
